Generalized scraping, but poorly...
Problem unifying data sources: different sources use different attribute names for the same type of data. We’re using a single database scheme so attributes for RSS parser and those for BeautifulSoup do not match. A mapping dictionary that standardizes the field names across sources is not sufficient since scraping uses HTML... For now, hardcoding a few attributes and not scraping everything