BIG CHANGE (switch to Weaviate)
Mostly refactored to generalize the option of using a different database. Isolated the database connection and LLM querying to a separate function to make things more agnostic.

The data in "results" is represented differently now, so as to consolidate both the Weaviate and ChromaDB methods.

Querying might break down if n_results > 1 but leaving for now as the use-case is not there (yet?).

The Local LLM is performing poorly now.. I do not think it's due to these changes!